work done on immediates

+++ Headers/getters.h
@@ -10,7 +10,7 @@ std::string get_opcode(std::string);
 bool is_alpha(char);
 bool is_num(char);
 std::string get_immediate_as_string(std::string);
-uint32_t get_immediate_as_int(std::string);
+uint16_t get_immediate_as_int(std::string);
 uint8_t get_opcode_as_int(std::string);
 std::vector<uint8_t> get_registers(std::string, int);
